+++ NEWS.md
@@ -1,5 +1,66 @@
 
-## reticulate 1.10 (CRAN)
+## reticulate 1.11 (CRAN)
+
+- Fixed an issue where attempts to activate virtual environments created with
+  virtualenv 16.4.1 would fail. (#437)
+
+- Fixed an issue where conversion of Pandas Categorical variables to R objects
+  would fail. (#389)
+
+- Textual output generated when adding items to a matplotlib plot object
+  are now suppressed.
+
+- If the last statement in a Python chunk returns a matplotlib plot object,
+  the plot will now be auto-shown as in other environments.
+
+- The reticulate function help handler now returns function arguments for
+  Python builtin functions.
+
+- Top-level Python statements can now include leading indent when submitted
+  with `repl_python()`.
+
+- The current `matplotlib` figure is now cleared as each Python chunk in an
+  R Markdown document is run.
+
+- The `r` helper object (used for evaluating R code from Python) now better
+  handles conversion of R functions. (#383)
+
+- The `use_virtualenv()` function now understands how to bind to virtual
+  environments created by the Python `venv` module.
+  
+- Reticulate better handles conversions of R lists to Python, and similarly,
+  Python lists to R. We now call `r_to_py()` on each sub-element of an R list,
+  and similarly, `py_to_r()` on each sub-element of a Python list.
+
+- Reticulate now always converts R `Date` objects into Python `datetime`
+  objects. Note that these conversions can be inefficient -- if you would
+  prefer conversion to NumPy `datetime64` objects / arrays, you should convert
+  your date to `POSIXct` first.
+
+- Python chunks containing errors will cause execution to halt if 'error=FALSE'
+  during render, conforming with the default knitr behavior for R chunks.
+
+- The output of bare statements (e.g. `1 + 1`) is now emitted as output when using
+  the reticulate Python engine.
+
+- Remapping of Python output streams to be R can now be explicitly enabled
+  by setting the environment variable `RETICULATE_REMAP_OUTPUT_STREAMS` to 1. (#335)
+
+- Allow syntax errors in Python chunks with 'eval = FALSE' (#343)
+
+- Avoid dropping blank lines in Python chunks (#328)
+
+- Use "agg" matplotlib backend when running under RStudio Desktop (avoids
+  crashes when attempting to generate Python plots)
+
+- Add `as.character()` S3 method for Python bytes (defaults to converting using 
+  UTF-8 encoding)
+  
+- Add `py_main_thread_func()` for providing R callbacks to Python libraries that may
+  invoke the function on a Python background thread.
+
+
+## reticulate 1.10
 
 - Output is now properly displayed when using the `reticulate` REPL with
   Windows + Python 2.7.